Initial implementation of Entry.
Added parameter constructors to Player, as well as a protected default constructor, to be used only by the spring framework.
Added EntryId, a class that represents the combined primary keys of Entry and required by spring.
Added placeholders for Game, Gamelist and Tag, as well as repositories for them.
Placeholder test files for Entry, Game, Tag and Gamelist.